+++ packages/woff2/woff2_ffi/tool/ffigen.dart
@@ -2,23 +2,15 @@ import 'dart:io';
 
 import 'package:ffigen/ffigen.dart';
 
-void main() {
+Future<void> main() async {
   final packageRoot = Platform.script.resolve("../");
-
-  final identifierRegExp = RegExp("^woff2_", caseSensitive: false);
-
-  bool include(Declaration declaration) =>
-      identifierRegExp.hasMatch(declaration.originalName);
-
   final resourceDir = _clangResourceDir();
 
-  FfiGenerator(
-    output: .new(
-      dartFile: packageRoot.resolve("lib/src/ffi_bindings.g.dart"),
-      format: true,
-      commentType: const .new(.any, .full),
-    ),
-    headers: .new(
+  final identifierRegExp = RegExp("^woff2", caseSensitive: false);
+  bool include(NamedNode node) => identifierRegExp.hasMatch(node.originalName);
+
+  await FfiGenerator(
+    input: .new(
       entryPoints: [
         // WOFF2 C API
         packageRoot.resolve("src/woff2_c_api.h"),
@@ -27,13 +19,29 @@ void main() {
         if (resourceDir != null) ...["-resource-dir", resourceDir],
       ],
     ),
-    functions: .new(include: include),
-    enums: .new(include: include),
-    structs: .new(include: include),
-    globals: .new(include: include),
-    macros: .new(include: include),
-    typedefs: .new(include: include),
-    unions: .new(include: include),
+    output: .new(
+      dart: .new(path: packageRoot.resolve("lib/src/ffi_bindings.g.dart")),
+      format: true,
+      commentType: const .new(.any, .full),
+    ),
+    visitors: [
+      .new(
+        func: (node) => node.isIncluded = include(node),
+        struct: (node) => node.isIncluded = include(node),
+        union: (node) => node.isIncluded = include(node),
+        enumClass: (node) => node.isIncluded = include(node),
+        global: (node) => node.isIncluded = include(node),
+        macroConstant: (node) => node.isIncluded = include(node),
+        typealias: (node) => node.isIncluded = include(node) ? .always : .never,
+        objCInterface: (node) => node.isIncluded = include(node),
+        objCProtocol: (node) => node.isIncluded = include(node),
+        objCCategory: (node) => node.isIncluded = include(node),
+        cppClass: (node) => node.isIncluded = include(node),
+        unnamedEnumConstant: (node) => node.isIncluded = include(node),
+        objCMethod: (node) => node.isIncluded = include(node),
+        cppMethod: (node) => node.isIncluded = include(node),
+      ),
+    ],
   ).generate();
 }